Conflicts:
	golem/network/p2p/p2pservice.py

+++ golem/testutils.py
@@ -6,6 +6,7 @@ from os import path, mkdir
 
 from golem.model import Database
 from golem.core.common import is_windows
+from golem.ethereum import Client
 
 
 class TempDirFixture(unittest.TestCase):
@@ -24,7 +25,12 @@ class TempDirFixture(unittest.TestCase):
         self.path = self.tempdir  # Alias for legacy tests
 
     def tearDown(self):
-        shutil.rmtree(self.tempdir)
+        # Firstly kill Ethereum node to clean up after it later on.
+        # FIXME: This is temporary solution. Ethereum node should always be
+        #        the explicit dependency and users should close it correctly.
+        Client._kill_node()
+        if path.isdir(self.tempdir):
+            shutil.rmtree(self.tempdir)
 
     def temp_file_name(self, name):
         return path.join(self.tempdir, name)
